+++ client/src/main/java/replicant/Subscription.java
@@ -170,7 +170,7 @@ public abstract class Subscription
   }
 
   @ObservableValueRef
-  protected abstract ObservableValue getEntitiesObservableValue();
+  protected abstract ObservableValue<?> getEntitiesObservableValue();
 
   @Override
   public int compareTo( @Nonnull final Subscription o )