+++ server/src/main/java/org/realityforge/replicant/server/ee/AbstractReplicationInterceptor.java
@@ -28,7 +28,7 @@ public abstract class AbstractReplicationInterceptor
     {
       return method.getDeclaringClass().getName() + "." + method.getName();
     }
-    final Constructor constructor = context.getConstructor();
+    final Constructor<?> constructor = context.getConstructor();
     if ( null != constructor )
     {
       return constructor.getDeclaringClass().getName() + "." + constructor.getName();