+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Upgrade the `org.realityforge.arez` artifacts to version `0.151`.
+* Suppress or eliminate rawtypes warnings and turn on linting to avoid their return.
 
 ### [v6.47](https://github.com/replicant4j/replicant/tree/v6.47) (2019-10-17) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.46...v6.47)